requirements.txt
# NBA Prediction System Dependencies

# Web Framework
streamlit>=1.30.0

# NBA Data API
nba_api>=1.4.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Visualization
plotly>=5.18.0

# HTTP Requests (for odds API)
requests>=2.31.0

# Utilities
python-dateutil>=2.8.2

# Database (Cloud)
psycopg2-binary>=2.9.9
